Hook NESTEST back up
nestest was not working because it was not jumping to the automated
start location, but reading that location as if it was the reset vector.

Add a `reset_override` method used to override the normal reset vector
for the NES, specifically for automated tests like nestest.

Also make the logging match up (not including the PPU or the arguments
to the emulated instructions, as these will be handled later).