Make nestest parse the log file
Logging to a file while running nestest in the same format as the
nestest log and diffing the two was hard to extend while work on the PPU
is being done.

Add a parser for the log that creates a struct, and compare it with the
CPU state. This will make it easier to keep the test passing while
adding the rest of CPU support.