PPU: Fetch tiles far enough in advance
The PPU does tile fetches two tiles in advance. Some games seem to start
setting up the palette far enough ahead that if this is not modeled the
new palette will be applied to the tile after the palette is changed,
and not two tiles after.

Add a VecDeque as a circular buffer for tiles which we cycle through.
Note that even though cycle 337 does have a nametable fetch, the buffer
should not be touched because these fetches are garbage.