Cartridge: Tidy cartridge implementation
Tidy up the cartridge implementation in preparation for a larger
refactor where we split the cartridge into parts needed by the different
components, to avoid needing the overhead of a shared cartridge. Each
component only accesses part of the cartridge, so this should be safe:
 - CPU: PRG
 - PPU: CHR
 - APU: DPCM

Tag points of interest with FIXMEs.