Remove RefCell from Cartridge
Completely remove RefCell from the Cartridge. As the PPU and APU do not
need access, the data required is simply cloned and passed to these
subsystems before being owned by the CPU.

If bank switching is required, the memory can be sent along with a remap
function as required, as the CPU and PPU control their own
bank-switching mechanism.