APU: Implement the DMC
The DMC is the delta modulation channel for the NES, playing delta
encoded samples from ROM. The output is a 7-bit value computed using
shift registers filled sequentially with each byte in a sample where
  * a 0 bit subtracts 2 from the current output
  * a 1 bit adds 2 to the current output

The sample start address and length are configured by SW where the
sample is played until the end is reached, at which point an IRQ may be
generated. Alternatively the sample may also be replayed using a loop
flag. The rate at which the output changes is also programmable,
specified in terms of CPU cycles