Optimize PPU
Fetching the tiles in a cycle-accurate manner was prohibitively
expensive. Instead, fetch them only when a new tile is being fetched
(every 8 PPU cycles).

Do not run the PPU unless the CPU changes its state such that it impacts
the frame, inspects its state, or an NMI will be generated during the
current frame.