Isolate shared state
Isolate state which will be shared between the interpeter and the binary
translator. Require all interpeter methods pass the state and mutate
that instead of having all state on a "god" struct. The interpreter is
what will communicate with the rest of the system, so this is where the
bus lives.