Only show timer results once
Previously a thread local was used for the registry to avoid
thread-safety concerns. This had the undesired side effect of printing
the timer durations multiple times. This also meant that results across
threads were not merged.

Instead, use a mutex to synchronize the initial population of the
TimeResultRegistry and atomics to ensure updates to timers are visible
across threads.