PPU: Box the FrameBuffer, have PPU wait on receive
The frame buffer is a little large now that it is tripple buffered. Box
it so it is on the heap, and only double-buffer it. Only two buffers
should be required, as it should not be possible for the PPU to advance
to write the inactive buffer until the renderer receives the next one:

         PPU                  renderer
push frame
process
                         receives command
push frame (blocks)
                         frame done
                         receives command